[MARCHETYPES-67] Fix maven-archetype-plugin test case
As described here https://issues.apache.org/jira/browse/MARCHETYPES-67

The POM test project for unit test was referring to a non existent
plugin "maven-my-plugin" instead of referring to the plugin created by
the artifact.

The unit test also did not check the correct configuration of the test
project. I've added a further assertion in the unit test; I made sure it
fails without the patch to the POM test project; with the patched POM
test project everything works.